+++ src/codemagic/cli/argument/__init__.py
@@ -5,5 +5,6 @@ from .argument_formatter import ArgumentFormatter
 from .argument_parser_builder import ArgumentParserBuilder
 from .argument_properties import ArgumentProperties
 from .common_argument_types import CommonArgumentTypes
+from .mutually_exclusive_groups import MutuallyExclusiveGroup
 from .typed_cli_argument import EnvironmentArgumentValue
 from .typed_cli_argument import TypedCliArgument