+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+Version UNRELEASED
+-------------
+
+**Development**
+- Add support for mutually exclusive groups on argument parser. [PR #422](https://github.com/codemagic-ci-cd/cli-tools/pull/422)
+
 Version 0.53.4
 -------------